+++ internal/admission/handler_test.go
@@ -16,6 +16,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	k8stypes "k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/annotations"
 	ctrlref "github.com/kong/kubernetes-ingress-controller/v3/internal/controllers/reference"
@@ -109,7 +110,10 @@ func TestHandleService(t *testing.T) {
 	tests := []struct {
 		name         string
 		resource     corev1.Service
+		validator    KongHTTPValidator
 		wantWarnings []string
+		wantMessage  string
+		isAllowed    bool
 	}{
 		{
 			name: "has kongingress annotation",
@@ -121,10 +125,12 @@ func TestHandleService(t *testing.T) {
 					},
 				},
 			},
+
 			wantWarnings: []string{
 				fmt.Sprintf(serviceWarning, annotations.AnnotationPrefix+annotations.ConfigurationKey,
 					kongv1beta1.KongUpstreamPolicyAnnotationKey),
 			},
+			isAllowed: true,
 		},
 		{
 			name: "has upstream policy annotation",
@@ -136,13 +142,97 @@ func TestHandleService(t *testing.T) {
 					},
 				},
 			},
-			wantWarnings: nil,
+			isAllowed: true,
+		},
+		{
+			name: "fails when many plugins of the same type are attached",
+			resource: corev1.Service{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "default",
+					Annotations: map[string]string{
+						annotations.AnnotationPrefix + annotations.PluginsKey: "plugin1, plugin2, plugin3",
+					},
+				},
+			},
+			validator: KongHTTPValidator{
+				ManagerClient: func() client.Client {
+					scheme := runtime.NewScheme()
+					require.NoError(t, kongv1.AddToScheme(scheme))
+					fakeClient := fake.NewClientBuilder().WithScheme(scheme).WithObjects(
+						&kongv1.KongPlugin{
+							ObjectMeta: metav1.ObjectMeta{
+								Name:      "plugin1",
+								Namespace: "default",
+							},
+							PluginName: "foo",
+						},
+						&kongv1.KongPlugin{
+							ObjectMeta: metav1.ObjectMeta{
+								Name:      "plugin2",
+								Namespace: "default",
+							},
+							PluginName: "bar",
+						},
+						&kongv1.KongPlugin{
+							ObjectMeta: metav1.ObjectMeta{
+								Name:      "plugin3",
+								Namespace: "default",
+							},
+							PluginName: "foo",
+						},
+					).Build()
+					return fakeClient
+				}(),
+			},
+			isAllowed:   false,
+			wantMessage: "Service has invalid KongPlugin annotation: cannot attach multiple plugins: plugin1, plugin3 of the same type foo",
+		},
+		{
+			name: "pass when many valid plugins are attached",
+			resource: corev1.Service{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "default",
+					Annotations: map[string]string{
+						annotations.AnnotationPrefix + annotations.PluginsKey:       "plugin1, plugin2, plugin3",
+						annotations.AnnotationPrefix + annotations.ConfigurationKey: "test",
+					},
+				},
+			},
+			validator: KongHTTPValidator{
+				ManagerClient: func() client.Client {
+					scheme := runtime.NewScheme()
+					require.NoError(t, kongv1.AddToScheme(scheme))
+					fakeClient := fake.NewClientBuilder().WithScheme(scheme).WithObjects(
+						&kongv1.KongPlugin{
+							ObjectMeta: metav1.ObjectMeta{
+								Name:      "plugin1",
+								Namespace: "default",
+							},
+							PluginName: "foo",
+						},
+						&kongv1.KongPlugin{
+							ObjectMeta: metav1.ObjectMeta{
+								Name:      "plugin2",
+								Namespace: "default",
+							},
+							PluginName: "bar",
+						},
+					).Build()
+					return fakeClient
+				}(),
+			},
+			isAllowed: true,
+			wantWarnings: []string{
+				fmt.Sprintf(serviceWarning, annotations.AnnotationPrefix+annotations.ConfigurationKey,
+					kongv1beta1.KongUpstreamPolicyAnnotationKey),
+			},
 		},
 	}
 	for _, tt := range tests {
 		resource := tt.resource
 		t.Run(tt.name, func(t *testing.T) {
-			validator := KongHTTPValidator{}
 			raw, err := json.Marshal(tt.resource)
 			require.NoError(t, err)
 			request := admissionv1.AdmissionRequest{
@@ -152,7 +242,7 @@ func TestHandleService(t *testing.T) {
 				},
 			}
 			handler := RequestHandler{
-				Validator: validator,
+				Validator: tt.validator,
 				Logger:    logr.Discard(),
 			}
 
@@ -160,8 +250,9 @@ func TestHandleService(t *testing.T) {
 
 			got, err := handler.handleService(context.Background(), request, responseBuilder)
 			require.NoError(t, err)
-			require.True(t, got.Allowed)
+			require.Equal(t, tt.isAllowed, got.Allowed)
 			require.Equal(t, tt.wantWarnings, got.Warnings)
+			require.Equal(t, tt.wantMessage, got.Result.Message)
 		})
 	}
 }